refactor(architecture): start merging modules: maven
